Split out utils into separate test file
