+++ packages/crypto/tests/utils-key-manager.spec.ts
@@ -1,37 +1,8 @@
 import { expect } from 'chai';
 
-import { bytesToBase58btcMultibase } from '../src/utils.js';
 import { checkValidProperty, checkRequiredProperty, uuid } from '../src/utils-key-manager.js';
 
 describe('Crypto Utils', () => {
-  describe('bytesToBase58btcMultibase()', () => {
-    it('returns a multibase encoded string', () => {
-    // Test Vector 1.
-      const input = {
-        header : new Uint8Array([0x00, 0x00]),
-        data   : new Uint8Array([0x00, 0x00])
-      };
-      const output = 'z1111';
-      const encoded = bytesToBase58btcMultibase(input.header, input.data);
-      expect(encoded).to.be.a.string;
-      expect(encoded.substring(0, 1)).to.equal('z');
-      expect(encoded).to.deep.equal(output);
-    });
-
-
-    it('returns multibase encoded value with specified header', () => {
-    // Test Vector 1.
-      const input = {
-        header : new Uint8Array([0x01, 0x02]),
-        data   : new Uint8Array([3, 4, 5, 6, 7])
-      };
-      const output = 'z3DUyZY2dc';
-
-      const encoded = bytesToBase58btcMultibase(input.header, input.data);
-      expect(encoded).to.deep.equal(output);
-    });
-  });
-
   describe('checkValidProperty()', () => {
     it('should not throw for a property in the allowed list', () => {
       expect(() => checkValidProperty({ property: 'foo', allowedProperties: ['foo', 'bar']})).to.not.throw();
